Move the rest of the history tests into test_history.py
